Profile_Screen and store the data by UID is done
the user is now be able to see own information in profile screen. when the user register in the registration page and now the user is able to choose and has 2 option to take picture by using camera and or choosing a photo the from mobile gallery. when the user filling the text-filed and choosing a picture  so  the registration information is also saved in Cloud Firestore in collection.